Enhancement: Renderer.ts
Define the createWindow function from the main.ts to seperate the
rendering settings from the main app settings.
